A GameCharacter can now carry multiple weapons and cycle through them.
To use, simply call "AddWeapon(Weapon weapon)" on a Game Character to add a Weapon to it. Of interest is the "AddWeaponAndSetActive" call that will make the character automatically use the weapon they just picked up.

The AddWeapon calls create add a copy of the Weapon component you pass onto the character.